requirements.txt
certifi==2019.9.11
chardet==3.0.4
clearbit==0.1.7
Django==2.2.5
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
idna==2.8
psycopg2==2.8.3
pyhunter==1.2
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6
